all: fix up names after move
