+++ macaroonbakery/checkers/_operation.py
@@ -1,6 +1,6 @@
 # Copyright 2017 Canonical Ltd.
 # Licensed under the LGPLv3, see LICENCE file for details.
-from macaroonbakery.checkers.auth_context import ContextKey
+from ._auth_context import ContextKey
 
 OP_KEY = ContextKey('op-key')
