+++ macaroonbakery/httpbakery/__init__.py
@@ -1,11 +1,11 @@
 # Copyright 2017 Canonical Ltd.
 # Licensed under the LGPLv3, see LICENCE file for details.
-from macaroonbakery.httpbakery.client import (
+from ._client import (
     BakeryException,
     Client,
     extract_macaroons,
 )
-from macaroonbakery.httpbakery.error import (
+from ._error import (
     BAKERY_PROTOCOL_HEADER,
     DischargeError,
     ERR_DISCHARGE_REQUIRED,
@@ -17,18 +17,18 @@ from macaroonbakery.httpbakery.error import (
     discharge_required_response,
     request_version,
 )
-from macaroonbakery.httpbakery.keyring import ThirdPartyLocator
-from macaroonbakery.httpbakery.interactor import (
+from ._keyring import ThirdPartyLocator
+from ._interactor import (
     DischargeToken,
     Interactor,
     LegacyInteractor,
     WEB_BROWSER_INTERACTION_KIND,
 )
-from macaroonbakery.httpbakery.browser import (
+from ._browser import (
     WebBrowserInteractionInfo,
     WebBrowserInteractor,
 )
-from macaroonbakery.httpbakery.discharge import discharge
+from ._discharge import discharge
 
 __all__ = [
     'BAKERY_PROTOCOL_HEADER',
