+++ tox.ini
@@ -22,7 +22,7 @@ commands =
 
 [testenv:lint]
 usedevelop = True
-commands = flake8 --ignore E501 --show-source macaroonbakery --exclude macaroonbakery/bakery/internal/id_pb2.py
+commands = flake8 --ignore E501 --show-source macaroonbakery --exclude macaroonbakery/bakery/_internal/id_pb2.py
 
 [testenv:docs]
 changedir = docs
